#407 update plugin config
